chore: extract shared parquet readers and unify factory loading
