+++ src/storage/parquet/eth_calls.rs
@@ -1,0 +1,188 @@
+//! Parquet readers for regular and event-triggered eth_call results.
+
+use std::fs::File;
+use std::path::Path;
+
+use arrow::array::{Array, BinaryArray, FixedSizeBinaryArray, UInt32Array, UInt64Array};
+use parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
+
+use super::ParquetReadError;
+use crate::decoding::{EthCallResult, EventCallResult};
+
+/// Read regular call results from a raw eth_call parquet file.
+///
+/// Each row contains: block_number, block_timestamp, address (20 bytes), value (binary).
+pub fn read_regular_calls_from_parquet(
+    path: &Path,
+) -> Result<Vec<EthCallResult>, ParquetReadError> {
+    let file = File::open(path)?;
+    let builder = ParquetRecordBatchReaderBuilder::try_new(file)?;
+    let reader = builder.build()?;
+
+    let mut results = Vec::new();
+
+    for batch_result in reader {
+        let batch = batch_result?;
+        let schema = batch.schema();
+
+        let block_number_idx = schema.index_of("block_number").ok();
+        let block_timestamp_idx = schema.index_of("block_timestamp").ok();
+        let address_idx = schema.index_of("address").ok();
+        let value_idx = schema.index_of("value").ok();
+
+        for row in 0..batch.num_rows() {
+            let block_number = block_number_idx
+                .and_then(|i| {
+                    batch
+                        .column(i)
+                        .as_any()
+                        .downcast_ref::<UInt64Array>()
+                        .map(|a| a.value(row))
+                })
+                .unwrap_or(0);
+
+            let block_timestamp = block_timestamp_idx
+                .and_then(|i| {
+                    batch
+                        .column(i)
+                        .as_any()
+                        .downcast_ref::<UInt64Array>()
+                        .map(|a| a.value(row))
+                })
+                .unwrap_or(0);
+
+            let contract_address = address_idx
+                .and_then(|i| {
+                    batch
+                        .column(i)
+                        .as_any()
+                        .downcast_ref::<FixedSizeBinaryArray>()
+                        .and_then(|a| {
+                            let bytes = a.value(row);
+                            if bytes.len() == 20 {
+                                let mut arr = [0u8; 20];
+                                arr.copy_from_slice(bytes);
+                                Some(arr)
+                            } else {
+                                None
+                            }
+                        })
+                })
+                .unwrap_or_default();
+
+            let value = value_idx
+                .and_then(|i| {
+                    batch
+                        .column(i)
+                        .as_any()
+                        .downcast_ref::<BinaryArray>()
+                        .map(|a| a.value(row).to_vec())
+                })
+                .unwrap_or_default();
+
+            results.push(EthCallResult {
+                block_number,
+                block_timestamp,
+                contract_address,
+                value,
+            });
+        }
+    }
+
+    Ok(results)
+}
+
+/// Read event-triggered call results from a raw eth_call parquet file.
+///
+/// Each row contains: block_number, block_timestamp, log_index, target_address (20 bytes), value (binary).
+pub fn read_event_calls_from_parquet(
+    path: &Path,
+) -> Result<Vec<EventCallResult>, ParquetReadError> {
+    let file = File::open(path)?;
+    let builder = ParquetRecordBatchReaderBuilder::try_new(file)?;
+    let reader = builder.build()?;
+
+    let mut results = Vec::new();
+
+    for batch_result in reader {
+        let batch = batch_result?;
+        let schema = batch.schema();
+
+        let block_number_idx = schema.index_of("block_number").ok();
+        let block_timestamp_idx = schema.index_of("block_timestamp").ok();
+        let log_index_idx = schema.index_of("log_index").ok();
+        let address_idx = schema.index_of("target_address").ok();
+        let value_idx = schema.index_of("value").ok();
+
+        for row in 0..batch.num_rows() {
+            let block_number = block_number_idx
+                .and_then(|i| {
+                    batch
+                        .column(i)
+                        .as_any()
+                        .downcast_ref::<UInt64Array>()
+                        .map(|a| a.value(row))
+                })
+                .unwrap_or(0);
+
+            let block_timestamp = block_timestamp_idx
+                .and_then(|i| {
+                    batch
+                        .column(i)
+                        .as_any()
+                        .downcast_ref::<UInt64Array>()
+                        .map(|a| a.value(row))
+                })
+                .unwrap_or(0);
+
+            let log_index = log_index_idx
+                .and_then(|i| {
+                    batch
+                        .column(i)
+                        .as_any()
+                        .downcast_ref::<UInt32Array>()
+                        .map(|a| a.value(row))
+                })
+                .unwrap_or(0);
+
+            let target_address = address_idx
+                .and_then(|i| {
+                    batch
+                        .column(i)
+                        .as_any()
+                        .downcast_ref::<FixedSizeBinaryArray>()
+                        .and_then(|a| {
+                            let bytes = a.value(row);
+                            if bytes.len() == 20 {
+                                let mut arr = [0u8; 20];
+                                arr.copy_from_slice(bytes);
+                                Some(arr)
+                            } else {
+                                None
+                            }
+                        })
+                })
+                .unwrap_or_default();
+
+            let value = value_idx
+                .and_then(|i| {
+                    batch
+                        .column(i)
+                        .as_any()
+                        .downcast_ref::<BinaryArray>()
+                        .map(|a| a.value(row).to_vec())
+                })
+                .unwrap_or_default();
+
+            results.push(EventCallResult {
+                block_number,
+                block_timestamp,
+                log_index,
+                target_address,
+                value,
+            });
+        }
+    }
+
+    Ok(results)
+}
