+++ src/storage/parquet/once_calls.rs
@@ -1,0 +1,109 @@
+//! Parquet reader for "once" call results.
+
+use std::collections::HashMap;
+use std::fs::File;
+use std::path::Path;
+
+use arrow::array::{Array, BinaryArray, FixedSizeBinaryArray, UInt64Array};
+use parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
+
+use super::ParquetReadError;
+use crate::decoding::eth_calls::CallDecodeConfig;
+use crate::decoding::OnceCallResult;
+
+/// Read "once" call results from a raw parquet file.
+///
+/// The `configs` parameter specifies which function columns to extract.
+/// Each row yields a `OnceCallResult` with a map of function_name -> raw result bytes.
+pub fn read_once_calls_from_parquet(
+    path: &Path,
+    configs: &[&CallDecodeConfig],
+) -> Result<Vec<OnceCallResult>, ParquetReadError> {
+    let file = File::open(path)?;
+    let builder = ParquetRecordBatchReaderBuilder::try_new(file)?;
+    let reader = builder.build()?;
+
+    let mut results = Vec::new();
+
+    for batch_result in reader {
+        let batch = batch_result?;
+        let schema = batch.schema();
+
+        let block_number_idx = schema.index_of("block_number").ok();
+        let block_timestamp_idx = schema.index_of("block_timestamp").ok();
+        let address_idx = schema
+            .index_of("address")
+            .or_else(|_| schema.index_of("address"))
+            .ok();
+
+        // Find function result columns
+        let function_indices: Vec<(String, Option<usize>)> = configs
+            .iter()
+            .map(|c| {
+                let col_name = format!("{}_result", c.function_name);
+                (c.function_name.clone(), schema.index_of(&col_name).ok())
+            })
+            .collect();
+
+        for row in 0..batch.num_rows() {
+            let block_number = block_number_idx
+                .and_then(|i| {
+                    batch
+                        .column(i)
+                        .as_any()
+                        .downcast_ref::<UInt64Array>()
+                        .map(|a| a.value(row))
+                })
+                .unwrap_or(0);
+
+            let block_timestamp = block_timestamp_idx
+                .and_then(|i| {
+                    batch
+                        .column(i)
+                        .as_any()
+                        .downcast_ref::<UInt64Array>()
+                        .map(|a| a.value(row))
+                })
+                .unwrap_or(0);
+
+            let contract_address = address_idx
+                .and_then(|i| {
+                    batch
+                        .column(i)
+                        .as_any()
+                        .downcast_ref::<FixedSizeBinaryArray>()
+                        .and_then(|a| {
+                            let bytes = a.value(row);
+                            if bytes.len() == 20 {
+                                let mut arr = [0u8; 20];
+                                arr.copy_from_slice(bytes);
+                                Some(arr)
+                            } else {
+                                None
+                            }
+                        })
+                })
+                .unwrap_or_default();
+
+            let mut function_results = HashMap::new();
+            for (function_name, idx_opt) in &function_indices {
+                if let Some(idx) = idx_opt {
+                    if let Some(arr) = batch.column(*idx).as_any().downcast_ref::<BinaryArray>() {
+                        if !arr.is_null(row) {
+                            function_results.insert(function_name.clone(), arr.value(row).to_vec());
+                        }
+                    }
+                }
+            }
+
+            results.push(OnceCallResult {
+                block_number,
+                block_timestamp,
+                contract_address,
+                results: function_results,
+            });
+        }
+    }
+
+    Ok(results)
+}
